+++ src/main/java/dev/sidgames/wmo/item/tool/WMOShovelItem.java
@@ -1,0 +1,12 @@
+package dev.sidgames.wmo.item.tool;
+
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemGroup;
+import net.minecraft.item.ShovelItem;
+import net.minecraft.item.ToolMaterial;
+
+public class WMOShovelItem extends ShovelItem {
+    public WMOShovelItem(ToolMaterial material, float attackDamage, float attackSpeed) {
+        super(material, attackDamage, attackSpeed, new Item.Settings().group(ItemGroup.TOOLS));
+    }
+}