+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added cobalt
 - Added zinc
+- Added enderite tools
 
 ## Updates
 
@@ -12,6 +13,9 @@
 ## Bug Fixes
 
 ## Technical Updates
+
+- Replace ShovelItem with custom WMOShovelItem class
+- Replace SwordItem with custom WMOSwordItem class
 
 # Overall changelog for 1.7.0